requirements.txt
numpy
pandas
scikit-learn
lightgbm
tensorflow
keras
flask
kafka-python
matplotlib
plotly
imbalanced-learn
tensorrt